+++ gen/drivers/query.go
@@ -285,15 +285,34 @@ func (q QueryArg) Types() []string {
 }
 
 func (c QueryCol) Type(currPkg string, i language.Importer, types Types) string {
+	return c.getType(currPkg, i, types, true)
+}
+
+func (c QueryCol) TypeNoImport(currPkg string, i language.Importer, types Types) string {
+	return c.getType(currPkg, i, types, false)
+}
+
+func (c QueryCol) getType(currPkg string, i language.Importer, types Types, doImport bool) string {
 	if c.Nullable == nil {
 		panic(fmt.Sprintf("Column %s has no nullable value defined", c.Name))
 	}
+	if !*c.Nullable {
+		typ, def := types.GetNameAndDef(currPkg, c.TypeName)
+		if doImport {
+			i.ImportList(def.Imports)
+		}
+		return typ
+	}
 
-	return types.GetNullable(currPkg, i, c.TypeName, *c.Nullable)
+	typ, imports := types.GetNullTypeWithImports(currPkg, c.TypeName)
+	if doImport {
+		i.ImportList(imports)
+	}
+	return typ.Name
 }
 
 func (c QueryArg) RandomExpr(currPkg string, i language.Importer, types Types) string {
-	typ := c.TypeDef(currPkg, i, types)
+	typ := c.TypeDef(currPkg, i, types, false)
 	var sb strings.Builder
 
 	if c.CanBeMultiple {
@@ -338,14 +357,14 @@ func (c QueryArg) RandomExpr(currPkg string, i language.Importer, types Types) s
 
 func (c QueryArg) Type(currPkg string, i language.Importer, types Types) string {
 	if c.CanBeMultiple {
-		return "[]" + c.TypeDef(currPkg, i, types)
+		return "[]" + c.TypeDef(currPkg, i, types, true)
 	}
-	return c.TypeDef(currPkg, i, types)
+	return c.TypeDef(currPkg, i, types, true)
 }
 
-func (c QueryArg) TypeDef(currPkg string, i language.Importer, types Types) string {
+func (c QueryArg) TypeDef(currPkg string, i language.Importer, types Types, doImport bool) string {
 	if len(c.Children) == 0 {
-		return c.Col.Type(currPkg, i, types)
+		return c.Col.getType(currPkg, i, types, doImport)
 	}
 
 	var sb strings.Builder